Rename light_on_off component to switch and update references in CMakeLists and source files
